+++ src/market.rs
@@ -1,11 +1,10 @@
 use async_trait::async_trait;
 use ethers::abi::{AbiDecode, AbiEncode};
-use ethers::prelude::rand::Rng;
 use ethers::prelude::*;
 use ethers::utils::keccak256;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::collections::HashMap;
+
 use std::error::Error;
 use std::fs;
 use std::time::SystemTime;
@@ -14,7 +13,6 @@ use tokio::time::{Duration, Instant};
 use tokio_stream::Stream;
 
 use ethers::types::Log;
-use tokio_stream::StreamExt;
 
 use crate::server;
 
@@ -1095,190 +1093,25 @@ async fn job_logs(
     Ok(Box::new(stream))
 }
 
+
 // --------------------------------------------------------------------------------------------------------------------------------------------------------
 //                                                                  TESTS
 // --------------------------------------------------------------------------------------------------------------------------------------------------------
 
 
-#[derive(Clone)]
-pub struct TestLogger {}
-
-#[async_trait]
-impl LogsProvider for TestLogger {
-    async fn new_jobs<'a>(
-        &'a self,
-        _client: &'a Provider<Ws>,
-    ) -> Result<Box<dyn Stream<Item = (H256, bool)> + 'a>, Box<dyn Error + Send + Sync>> {
-        let logs: Vec<Log> = Vec::new();
-        Ok(Box::new(
-            tokio_stream::iter(
-                logs.iter()
-                    .map(|job| (job.topics[1], false))
-                    .collect::<Vec<_>>(),
-            )
-            .throttle(Duration::from_secs(2)),
-        ))
-    }
-
-    async fn job_logs<'a>(
-        &'a self,
-        _client: &'a Provider<Ws>,
-        job: H256,
-    ) -> Result<Box<dyn Stream<Item = Log> + Send + 'a>, Box<dyn Error + Send + Sync>> {
-        let logs: Vec<Log> = Vec::new();
-        Ok(Box::new(
-            tokio_stream::iter(
-                logs.into_iter()
-                .filter(|log| log.topics[1] == job)
-                .collect::<Vec<_>>())
-                .throttle(Duration::from_secs(2)),
-        ))
-    }
-}
-
-#[derive(Clone, Debug)]
-struct SpinUpOutcome {
-    time: Instant,
-    job: String,
-    instance_type: String,
-    region: String,
-    req_mem: i64,
-    req_vcpu: i32,
-    bandwidth: u64,
-    eif_url: String,
-}
-
-#[derive(Clone, Debug)]
-struct SpinDownOutcome {
-    time: Instant,
-    job: String,
-    _instance_id: String,
-    region: String,
-}
-
-#[derive(Clone, Debug)]
-enum TestAwsOutcome {
-    SpinUp(SpinUpOutcome),
-    SpinDown(SpinDownOutcome),
-}
-
-#[derive(Clone, Default)]
-pub struct TestAws {
-    outcomes: Vec<TestAwsOutcome>,
-    instances: HashMap<String, String>,
-}
-
-#[async_trait]
-impl InfraProvider for TestAws {
-    async fn spin_up(
-        &mut self,
-        eif_url: &str,
-        job: String,
-        instance_type: &str,
-        region: String,
-        req_mem: i64,
-        req_vcpu: i32,
-        bandwidth: u64,
-    ) -> Result<String, Box<dyn Error + Send + Sync>> {
-        self.outcomes.push(TestAwsOutcome::SpinUp(SpinUpOutcome {
-            time: Instant::now(),
-            job: job.clone(),
-            instance_type: instance_type.to_owned(),
-            region,
-            req_mem,
-            req_vcpu,
-            bandwidth,
-            eif_url: eif_url.to_owned(),
-        }));
-
-        let res = self.instances.get_key_value(&job);
-        if let Some(x) = res {
-            return Ok(x.1.clone());
-        }
-
-        let id: String = rand::thread_rng()
-            .sample_iter(rand::distributions::Alphanumeric)
-            .take(8)
-            .map(char::from)
-            .collect();
-        self.instances.insert(job, id.clone());
-
-        Ok(id)
-    }
-
-    async fn spin_down(
-        &mut self,
-        instance_id: &str,
-        job: String,
-        region: String,
-    ) -> Result<bool, Box<dyn Error + Send + Sync>> {
-        self.outcomes
-            .push(TestAwsOutcome::SpinDown(SpinDownOutcome {
-                time: Instant::now(),
-                job: job.clone(),
-                _instance_id: instance_id.to_owned(),
-                region,
-            }));
-
-        self.instances.remove(&job);
-
-        Ok(true)
-    }
-
-    async fn get_job_instance(
-        &mut self,
-        job: &str,
-        _region: String,
-    ) -> Result<(bool, String, String), Box<dyn Error + Send + Sync>> {
-        let res = self.instances.get_key_value(job);
-        if let Some(x) = res {
-            return Ok((true, x.1.clone(), "running".to_owned()));
-        }
-
-        Ok((false, String::new(), String::new()))
-    }
-
-    async fn check_instance_running(
-        &mut self,
-        _instance_id: &str,
-        _region: String,
-    ) -> Result<bool, Box<dyn Error + Send + Sync>> {
-        // println!("TEST: check_instance_running | instance_id: {}, region: {}", instance_id, region);
-        Ok(true)
-    }
-
-    async fn check_enclave_running(
-        &mut self,
-        _instance_id: &str,
-        _region: String,
-    ) -> Result<bool, Box<dyn Error + Send + Sync>> {
-        Ok(true)
-    }
-
-    async fn run_enclave(
-        &mut self,
-        _job: String,
-        _instance_id: &str,
-        _region: String,
-        _image_url: &str,
-        _req_vcpu: i32,
-        _req_mem: i64,
-        _bandwidth: u64,
-    ) -> Result<(), Box<dyn Error + Send + Sync>> {
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::market;
-    use crate::market::TestAwsOutcome;
-    use crate::test::{self, Action};
     use ethers::abi::AbiEncode;
     use ethers::prelude::*;
     use std::str::FromStr;
     use tokio::time::{sleep, Duration, Instant};
 
+    use crate::market;
+    use crate::test::{self, 
+        Action, 
+        TestAwsOutcome, 
+        TestAws
+    };
 
     #[tokio::test(start_paused = true)]
     async fn test_instance_launch_after_delay_on_spin_up() {
@@ -1297,7 +1130,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1361,7 +1194,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1426,7 +1259,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1487,7 +1320,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1524,7 +1357,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1561,7 +1394,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1598,7 +1431,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1635,7 +1468,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1672,7 +1505,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1709,7 +1542,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1747,7 +1580,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1811,7 +1644,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1872,7 +1705,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1935,7 +1768,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -1974,7 +1807,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
@@ -2013,7 +1846,7 @@ mod tests {
                 log
             })
             .chain(tokio_stream::pending()));
-        let mut aws: market::TestAws = Default::default();
+        let mut aws: TestAws = Default::default();
         let res = market::job_manager_once(
             job_stream,
             &mut aws,
